More efficient
Now, before the program walks to a given area it will check to see if the random location is minable ground based on RGB color value. This should help with being able to select a larger area of the screen and the program will find a location that will make you money. This will also prevent it from clicking on other players as often thus delating finding a new location longer.